Fix workflow controller test
The tests presents multiple validation errors to the controller.  The
last change to the controller reordered some validation steps, resulting
in a different (yet correct) response code from the controller.  This
commit modifies the test so that the first failed validation is the one
that is meant to be tested.